Moved metrics to common ingest part
