+++ pkg/pipeline/ingest/ingest_collector.go
@@ -27,14 +27,12 @@ import (
 	ms "github.com/mitchellh/mapstructure"
 	"github.com/netobserv/flowlogs-pipeline/pkg/api"
 	"github.com/netobserv/flowlogs-pipeline/pkg/config"
-	operationalMetrics "github.com/netobserv/flowlogs-pipeline/pkg/operational/metrics"
 	pUtils "github.com/netobserv/flowlogs-pipeline/pkg/pipeline/utils"
 	goflowFormat "github.com/netsampler/goflow2/format"
 	goflowCommonFormat "github.com/netsampler/goflow2/format/common"
 	_ "github.com/netsampler/goflow2/format/protobuf"
 	goflowpb "github.com/netsampler/goflow2/pb"
 	"github.com/netsampler/goflow2/utils"
-	"github.com/prometheus/client_golang/prometheus"
 	log "github.com/sirupsen/logrus"
 	"google.golang.org/protobuf/proto"
 )
@@ -59,16 +57,6 @@ type ingestCollector struct {
 type TransportWrapper struct {
 	c chan map[string]interface{}
 }
-
-var queueLength = operationalMetrics.NewGauge(prometheus.GaugeOpts{
-	Name: "ingest_collector_queue_length",
-	Help: "Queue length",
-})
-
-var linesProcessed = operationalMetrics.NewCounter(prometheus.CounterOpts{
-	Name: "ingest_collector_flow_logs_processed",
-	Help: "Number of log lines (flow logs) processed",
-})
 
 func NewWrapper(c chan map[string]interface{}) *TransportWrapper {
 	tw := TransportWrapper{c: c}
